change | upgradedata에 일부값을 float로 넣을 수 있도록 변경

+++ Cinema_Manager/Assets/02.Scripts/01.Managers/UpgradeManager.cs
@@ -3,11 +3,11 @@ using TMPro;
 using UnityEngine;
 
 public class UpgradeManager : MonoBehaviour {
-    public int[] playerMoveSpeedStat = new int[5];
+    public float[] playerMoveSpeedStat = new float[5];
     public int[] playerVolumeStat = new int[5];
     public float[] playerSellingcostStat = new float[5];
 
-    public int[] employeeMoveSpeedStat = new int[5];
+    public float[] employeeMoveSpeedStat = new float[5];
     public int[] employeeVolumeStat = new int[5];
     public int[] employeeAddStat = new int[5];
 
@@ -55,14 +55,14 @@ public class UpgradeManager : MonoBehaviour {
     }
     private void SettingPlayerStat() {
         float costWeight = playerSellingcostStat[_gameData.p_sellingcostLevel];
-        int speed = playerMoveSpeedStat[_gameData.p_movespeedLevel];
+        float speed = playerMoveSpeedStat[_gameData.p_movespeedLevel];
         int stack = playerVolumeStat[_gameData.p_volumeLevel];
 
         _player.SetPlayerStat(costWeight, speed, stack);
     }
 
     private void SettingEmployeeStat() {
-        int speed = employeeMoveSpeedStat[_gameData.e_movespeedLevel];
+        float speed = employeeMoveSpeedStat[_gameData.e_movespeedLevel];
         int stack = employeeVolumeStat[_gameData.e_volumeLevel];
         int employeeCount = employeeAddStat[_gameData.e_employmentLevel];
